Moved CRUD tests to model test folder as to keep structure during added tests.
