+++ src/test/java/com/javafullstacklibrary/model/ItemM2MCRUDTests.java
@@ -1,6 +1,5 @@
-package com.javafullstacklibrary;
+package com.javafullstacklibrary.model;
 
-import com.javafullstacklibrary.model.*;
 import org.junit.jupiter.api.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
